Rename Wish to Operation
